refactor(provider-models): move add and pull flows to pages
