+++ src/app/settings/provider/[providerId]/model-add.tsx
@@ -1,0 +1,1 @@
+export { ProviderModelAddScreen as default } from '@/screens/SettingsScreen/ProviderScreen';
