+++ src/app/settings/provider/[providerId]/model-pull.tsx
@@ -1,0 +1,1 @@
+export { ProviderModelPullScreen as default } from '@/screens/SettingsScreen/ProviderScreen';
