+++ src/screens/SettingsScreen/ProviderScreen/index.ts
@@ -4,3 +4,5 @@ export { useProviderDetailSettings } from './detail';
 export { default as ProviderApiKeySettingsScreen } from './ProviderApiKeySettingsScreen';
 export { default as ProviderDetailScreen } from './ProviderDetailScreen';
 export { default as ProviderEndpointSettingsScreen } from './ProviderEndpointSettingsScreen';
+export { default as ProviderModelAddScreen } from './ProviderModelAddScreen';
+export { default as ProviderModelPullScreen } from './ProviderModelPullScreen';
